refactor(ai): split tool definitions and executors into feature-scoped subfolder files
Create constants/toolDefinitions/ and services/toolExecutors/ subfolders
with one file per feature group. Master files become thin aggregators.

Tool definitions split into:
- toolDefinitions/entryToolDefinitions.ts       (log entry CRUD + macros)
- toolDefinitions/mealPlanToolDefinitions.ts    (meal plan)
- toolDefinitions/templateToolDefinitions.ts    (search + food/recipe templates)
- toolDefinitions/memoryToolDefinitions.ts      (memory)

Tool executors split into:
- toolExecutors/entryToolExecutors.ts           (log entry CRUD + macros)
- toolExecutors/mealPlanToolExecutors.ts        (meal plan)
- toolExecutors/templateToolExecutors.ts        (search + food/recipe templates)
- toolExecutors/memoryToolExecutors.ts          (memory)

Old flat files constants/templateToolDefinitions.ts and
services/templateToolExecutors.ts are removed.

Tool renames for clarity:
- read_entries         → read_log_entries
- create_entry         → log_food
- move_entry           → move_log_entry
- update_entry         → update_log_entry
- remove_entry         → delete_log_entry
- read_recent_entries  → read_recent_log_entries
- search_templates     → search_library

Cross-reference descriptions updated to reflect new tool names.
